+++ appinfo/routes.php
@@ -8,5 +8,6 @@ return [
         ['name' => 'admin#apply', 'url' => '/apply', 'verb' => 'POST'],
         ['name' => 'admin#getLogLevel', 'url' => '/log-level', 'verb' => 'GET'],
         ['name' => 'admin#setLogLevel', 'url' => '/log-level', 'verb' => 'POST'],
+        ['name' => 'admin#debugAllinfo', 'url' => '/debug-allinfo', 'verb' => 'POST'],
     ],
 ];